requirements.txt
asgiref==3.4.1
Django==3.2.9
djangorestframework==3.12.4
djangorestframework-simplejwt==5.1.0
sqlparse==0.5.3
django-environ==0.12.0
django-simpleui==2021.11.5
django-import-export==2.6.1
django-cors-headers==3.10.0
django-celery-beat==2.2.1
django-celery-results==2.2.0
django-extensions==3.1.5
django-health-check==3.17.0
django-filter==21.1
django-anymail==8.6
django-redis==5.0.0
drf-yasg==1.20.0
psycopg2-binary==2.9.10
flake8==7.1.2
isort==6.0.0
autopep8==2.3.2
celery==5.2.1
eventlet==0.33.1
dnspython==2.2.1
cryptography==42.0.7
hyperliquid-python-sdk==0.10.1
eth-account==0.13.5
redlock-py==1.0.8
pydantic==2.10.6
pika~=1.3.2